added custom color and font
