+++ src/components/home/categories-section/CategoryItem.tsx
