+++ src/components/home/categories-section/CategorySlide.tsx
