+++ src/components/home/hero-section/Searchbar/ClearDataButton.tsx
