+++ src/components/home/hero-section/Searchbar/StayLocation.tsx
